excel-llm-powerquery-final/requirements.txt
fastmcp
openai
python-dotenv
pywin32
